Fixed null materials not being transparent. Fixed some compiler warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2859 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ networking/url.cpp
@@ -129,7 +129,7 @@ const URL URL::getDir() const//get everthing except the file at the end
 	{
 		if(file[i] == '/' || file[i] == '\\')
 		{
-			assert(i+1 <= file.length());
+			assert(i+1 <= (int)file.length());
 
 			return URL(host, file.substr(0, i+1));
 		}
@@ -167,7 +167,7 @@ const URL URL::getOneDirUp() const
 	{
 		if(curdir.getFile()[i] == '/' || curdir.getFile()[i] == '\\')
 		{
-			assert(i+1 <= curdir.getFile().length());
+			assert(i+1 <= (int)curdir.getFile().length());
 
 			return URL(curdir.getHost(), curdir.getFile().substr(0, i+1));
 		}
